requirements.txt
pymatgen
mp-api
python-dotenv
pandas
pyarrow
google-cloud-storage